requirements.txt
numpy
pyyaml
scikit-image
tensorboardX
torch
torchvision